chore: add MIT License (#13)
Add MIT License to ensure proper licensing and legal compliance. This
change addresses the licensing requirements for the project and provides
clear permissions to users.

The MIT License is a permissive open-source license widely used for
software projects. It allows users to freely use, modify, and distribute
the software, either in its original form or with modifications, while
providing a disclaimer of warranty. This license is known for its
simplicity and flexibility, making it a popular choice for many
open-source projects.

Signed-off-by: Lapotor <17144397+Lapotor@users.noreply.github.com>